Improve documentation and naming.
