+++ examples/smallest_example_.py
@@ -5,7 +5,7 @@ import numpy as np
 from numpy.linalg import norm
 from scipy.sparse.linalg import gmres
 
-from htool_test import HMatrix
+from htool_ import HMatrix
 
 # SETUP
 # n² points on a regular grid in a square
@@ -15,16 +15,16 @@ for j in range(0, n):
     for k in range(0, n):
         points[j+k*n, :] = (j, k, 1)
 
+# BUILDING THE H-MATRIX
 def get_coef(i, j, coef):
     coef[0] = 1.0 / (1e-5 + norm(points[i, :] - points[j, :]))
 
-def get_submatrix(I, J, n, m, coef):
-    for i in range(0,n):
-        for j in range(0,m):
-            coef[j*n+i] = 1.0 / (1.e-5 + norm(points[I[i], :] - points[J[j], :]))
-
-# BUILDING THE H-MATRIX
 H = HMatrix.from_coefs(points, get_coef, epsilon=1e-3, eta=100, minclustersize=10)
+
+# def get_submatrix(I, J, n, m, coef):
+#     for i in range(0,n):
+#         for j in range(0,m):
+#             coef[j*n+i] = 1.0 / (1.e-5 + norm(points[I[i], :] - points[J[j], :]))
 # H = HMatrix.from_submatrices(points, get_submatrix, epsilon=1e-3, eta=100, minclustersize=10)
 
 print("Shape:", H.shape)
@@ -38,10 +38,10 @@ full_H = 1.0 / (1e-5 + norm(points.reshape(1, n*n, 3) - points.reshape(n*n, 1, 3
 y = np.ones((n*n,))
 
 x, _ = gmres(H, y)
-print("Error from gmres:", norm(H @ x - y))
+print("Error from gmres (Hmatrix):", norm(H @ x - y))
 
 x_full, _ = gmres(full_H, y)
-print("Error from gmres:", norm(full_H @ x_full - y))
+print("Error from gmres (full matrix):", norm(full_H @ x_full - y))
 
-print(norm(x - x_full)/norm(x))
+print("Error between the two solutions:", norm(x - x_full)/norm(x))
